add newEnemy & fixed a bug with keyInput

+++ Meteors/src/ua/kas/main/HUD.java
@@ -5,9 +5,10 @@ import java.awt.Graphics;
 
 public class HUD {
 
-	public static int HEALTH = 100;
+	public static float HEALTH = 100;
 
-	private int greenValue = 255;
+	private float greenValue = 255;
+
 	private int score = 0;
 	private int level = 1;
 
@@ -21,8 +22,8 @@ public class HUD {
 	public void render(Graphics g) {
 		g.setColor(Color.GRAY);
 		g.fillRect(15, 15, 200, 32);
-		g.setColor(new Color(75, greenValue, 0));
-		g.fillRect(15, 15, HEALTH * 2, 32);
+		g.setColor(new Color(75, (int) greenValue, 0));
+		g.fillRect(15, 15, (int) (HEALTH * 2), 32);
 		g.setColor(Color.WHITE);
 		g.drawRect(15, 15, 200, 32);
